+++ CHANGELOG.md
@@ -16,7 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add prettier plugin for beautiful code [(35)]
 - Add Confirmed payment attempt, used for advance[(37)]
 - fix Confirmed payment attempt response[(38)]
-- add Bank in attempt[()]
+- add Bank in attempt[(39)]
+fix Confirmed payment attempt with multi Confirmed payment [(40)]
 
 ### Changed
 - Change input of asssignRepresentativeToPayerEnterprise service [(29)]
@@ -113,4 +114,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [(33)]: https://bitbucket.org/cumplo/srm-apollo-server/pull-requests/33/
 [(37)]: https://bitbucket.org/cumplo/srm-apollo-server/pull-requests/37/
 [(38)]: https://bitbucket.org/cumplo/srm-apollo-server/pull-requests/38/
-[()]: https://bitbucket.org/cumplo/srm-apollo-server/pull-requests//
+[(39)]: https://bitbucket.org/cumplo/srm-apollo-server/pull-requests/39/
+[(40)]: https://bitbucket.org/cumplo/srm-apollo-server/pull-requests/40/